backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
playwright==1.58.0
beautifulsoup4==4.14.3
google-genai==1.72.0
pydantic==2.12.5
python-multipart==0.0.26
httpx==0.28.1
pillow==12.2.0
lxml==6.0.3
